requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-cors-headers
python-dotenv